cart

+++ src/main/java/com/buliang/pojo/OrderDetail.java
@@ -12,7 +12,7 @@ public class OrderDetail implements Serializable{
 	private Integer id;//ID
 	private Integer orderId;//订单ID
 	private Integer quantity;//数量
-	private Float cost;//单价
+	private Double cost;//单价
 	private Integer productId;//商品id
 
 	private Product product;//商品
